Phylo

Package for creating and manipulating phylogenies

Phylo is a Julia package that provides functionality for generating phylogenetic trees to feed into our Diversity package to calculate phylogenetic diversity (currently on master, but not released). Both are currently under development, so please raise an issue if you find any problems. Currently the package can be used to make trees manually, and to generate random trees using the framework from Distributions. For instance, to construct a sampler for 5 tip non-ultrametric trees, and then generate one or two random tree of that type:

julia> using Phylo

julia> nu = Nonultrametric(5);

julia> tree = rand(nu)
RootedTree with 5 tips, 9 nodes and 8 branches.
Leaf names are tip 3, tip 5, tip 2, tip 4 and tip 1

julia> trees = rand(nu, ["Tree 1", "Tree 2"])
TreeSet with 2 trees, each with 5 tips.
Tree names are Tree 2 and Tree 1

Tree 1: RootedTree with 5 tips, 9 nodes and 8 branches.
Leaf names are tip 5, tip 3, tip 4, tip 2 and tip 1

Tree 2: RootedTree with 5 tips, 9 nodes and 8 branches.
Leaf names are tip 3, tip 5, tip 1, tip 2 and tip 4

The code also provides iterators, and filtered iterators over the branches, nodes, branchnames and nodenames of a tree:

julia> collect(nodeiter(tree))
9-element Vector{LinkNode{OneRoot, String, Dict{String, Any}, LinkBranch{OneRoot, String, Dict{String, Any}, Float64}}}:
 LinkNode tip 1, a tip of the tree with an incoming connection (branch 16).

 LinkNode tip 2, a tip of the tree with an incoming connection (branch 10).

 LinkNode tip 3, a tip of the tree with an incoming connection (branch 11).

 LinkNode tip 4, a tip of the tree with an incoming connection (branch 14).

 LinkNode tip 5, a tip of the tree with an incoming connection (branch 9).

 LinkNode Node 6, an internal node with 1 inbound and 2 outbound connections (branches 12 and 9, 10)

 LinkNode Node 7, an internal node with 1 inbound and 2 outbound connections (branches 13 and 11, 12)

 LinkNode Node 8, an internal node with 1 inbound and 2 outbound connections (branches 15 and 13, 14)

 LinkNode Node 9, a root node with 2 outbound connections (branches 15, 16)

julia> collect(nodenamefilter(isroot, tree))
1-element Vector{String}:
 "Node 9"

TreeSets are iterators themselves

julia> collect(trees)

2-element Vector{RootedTree}:
 RootedTree with 5 tips, 9 nodes and 8 branches.
Leaf names are tip 3, tip 5, tip 1, tip 2 and tip 4

 RootedTree with 5 tips, 9 nodes and 8 branches.
Leaf names are tip 3, tip 5, tip 1, tip 2 and tip 4

The current main purpose of this package is to provide a framework for phylogenetics to use in our [Diversity][diversity-url] package, and they will both be adapted as appropriate until both are functioning as required (though they are currently working together reasonably successfully).

It can also read newick trees either from strings or files:

julia> using Phylo

julia> simpletree = parsenewick("((,Tip:1.0)Internal,)Root;")
RootedTree with 3 tips, 5 nodes and 4 branches.
Leaf names are Tip, Node 1 and Node 4

julia> getbranches(simpletree)
skipmissing(Union{Missing, LinkBranch{OneRoot, String, Dict{String, Any}, Float64}}[LinkBranch 1, from node Internal to node Tip (length 1.0).
, LinkBranch 2, from node Internal to node Node 1.
, LinkBranch 3, from node Root to node Internal.
, LinkBranch 4, from node Root to node Node 4.
])

julia> tree = open(parsenewick, Phylo.path("H1N1.newick"))
RootedTree with 507 tips, 1013 nodes and 1012 branches.
Leaf names are 227, 294, 295, 110, 390, ... [501 omitted] ... and 418

And it can read nexus trees from files too:

jjulia> ts = open(parsenexus, Phylo.path("H1N1.trees"))
[ Info: Created a tree called "TREE1"
[ Info: Created a tree called "TREE2"
TreeSet with 2 trees, each with 507 tips.
Tree names are TREE2 and TREE1

TREE1: RootedTree with 507 tips, 1013 nodes and 1012 branches.
Leaf names are H1N1_A_BRAZIL_11_1978, H1N1_A_TAHITI_8_1998, H1N1_A_TAIWAN_1_1986, H1N1_A_BAYERN_7_1995, H1N1_A_ENGLAND_45_1998, ... [501 omitted] ... and H1N1_A_PUERTORICO_8_1934

TREE2: RootedTree with 507 tips, 1013 nodes and 1012 branches.
Leaf names are H1N1_A_BRAZIL_11_1978, H1N1_A_TAHITI_8_1998, H1N1_A_TAIWAN_1_1986, H1N1_A_BAYERN_7_1995, H1N1_A_ENGLAND_45_1998, ... [501 omitted] ... and H1N1_A_PUERTORICO_8_1934

julia> ts["TREE1"]
RootedTree with 507 tips, 1013 nodes and 1012 branches.
Leaf names are H1N1_A_BRAZIL_11_1978, H1N1_A_TAHITI_8_1998, H1N1_A_TAIWAN_1_1986, H1N1_A_BAYERN_7_1995, H1N1_A_ENGLAND_45_1998, ... [501 omitted] ... and H1N1_A_PUERTORICO_8_1934

julia> gettreeinfo(ts)
Dict{String, Dict{String, Any}} with 2 entries:
  "TREE2" => Dict("lnP"=>-1.0)
  "TREE1" => Dict("lnP"=>1.0)

And while we wait for me (or kind contributors!) to fill out the other extensive functionality that many phylogenetics packages have in other languages, the other important feature that it offers is a fully(?)-functional interface to R, allowing any existing R library functions to be carried out on julia trees, and trees to be read from disk and written using R helper functions. Naturally the medium-term plan is to fill in as many of these gaps as possible in Julia, and as a result this R interface is not built into the package as it will make RCall (and R) a dependency, which I wanted to avoid. Instead, if you want to use the R interface you need to do it manually, as below:

julia> using RCall

julia> include(Phylo.path("rcall.jl", dir = "src"));

R> library(ape)

You can then translate back and forth using rcopy on R phylo objects, and RObject constructors on julia NamedTree types to keep them in Julia or @rput to move the object into R:

julia> rt = rcall(:rtree, 10)
RCall.RObject{RCall.VecSxp}

Phylogenetic tree with 10 tips and 9 internal nodes.

Tip labels:
	t10, t8, t1, t2, t6, t5, ...

Rooted; includes branch lengths.

julia> jt = rcopy(NamedTree, rt)
NamedTree with 10 tips, 19 nodes and 18 branches.
Leaf names are t8, t3, t7, t9, t6, ... [4 omitted] ... and t1

julia> rjt = RObject(jt); # manually translate it back to R

R> if (all.equal($rjt, $rt)) "no damage in translation"
[1] "no damage in translation"

julia> @rput rt; # Or use macros to pass R object back to R

julia> @rput jt; # And automatically translate jt back to R

R> jt

Phylogenetic tree with 10 tips and 9 internal nodes.

Tip labels:
	t10, t8, t1, t2, t6, t5, ...

Rooted; includes branch lengths.

R> if (all.equal(rt, jt)) "no damage in translation"
[1] "no damage in translation"

Phylo.API

Phylo

Phylo.Phylo — Module

Phylo package

The Phylo package provides some simple phylogenetics types (e.g. NamedTree) to interface to the Diversity package for measuring phylogenetic diversity. It also provides an interface to R for copying trees to and from that language and can read newick and nexus tree files (including TreeSets that contain multiple trees).

Finally it also provides a standard abstract interface to phylogenetic trees, by defining AbstractNode, AbstractBranch and AbstractTree supertypes, and methods to interface to them. It also provides (through the Phylo.API submodule) methods to (re)define to write your own phylogenetic type in a way that will interact cleanly with other phylogenetic packages.

Phylo.BinaryNode — Type

BinaryNode{B}(AbstractVector{B}, AbstractVector{B}) <: AbstractNode

A node of strict binary phylogenetic tree

Phylo.BinaryTree — Type

BinaryTree

Binary phylogenetic tree object with known leaves and per node data

Phylo.Branch — Type

Branch

A branch connecting two AbstractNodes of a phylogenetic tree

Phylo.BrownianTrait — Type

BrownianTrait{T <: AbstractTree, N <: Number}

A continuous trait evolved on a phylogenetic tree. This is a Sampleable type, so a random trait can be created using rand(). The trait to be evolved can be any continuous numeric type, including Unitful types for instance, and in the simplest case is determined by the third argument to the constructor start:

function BrownianTrait(tree::AbstractTree, trait::String, start::Number = 0.0; σ² = missing, σ = missing, f::Function = identity)

Note that when Unitful is being used, either here or in branch lengths, σ/σ² keyword argument units must be appropriate. The final keyword argument, f, is a function to transform the evolved gaussian trait into its true value. By default this is the identity function, but can, for instance, be abs to force a positive value on the trait, or more complex functions as required, such as a transformation to turn a continuous variable into a discrete trait

Phylo.DiscreteTrait — Type

DiscreteTrait{T <: AbstractTree, E <: Enum}

A discrete trait evolved on a phylogenetic tree. This is a Sampleable type, so a random trait can be created using rand(dt). The trait to be evolved must be an Enum (generally created using @enum), and is the second argument to the constructor:

function DiscreteTrait(tree::AbstractTree, ttype::Type{<:Enum}, transition_matrix::AbstractMatrix{Float64}, trait::String = "ttype")

The transition matrix holds transition rates from row to column (so row sums must be zero), and the transition probabilities in a branch are calculated as exp(transition_matrix .* branch_length).

Phylo.NamedBinaryTree — Type

NamedBinaryTree

Binary phylogenetic tree object with known leaves

Phylo.NamedTree — Type

NamedPolytomousTree

Polytomous phylogenetic tree object with known leaves

Phylo.Node — Type

Node{RT, NL, T}(AbstractVector{T}, AbstractVector{T}) <: AbstractNode

A node of potentially polytomous phylogenetic tree

Phylo.Nonultrametric — Type

Nonultrametric{T <: AbstractTree,
               SAMP <: Sampleable}(n::Int,
                                   sampleable::SAMP = Exponential())
Nonultrametric{T <: AbstractTree,
               SAMP <: Sampleable}(tiplabels::Vector{String},
                                   sampleable::SAMP = Exponential())

The sampler for non-ultrametric phylogenetic trees of size n or with tip labels tiplabels. Generate random trees by calling rand().

Phylo.PolytomousTree — Type

PolytomousTree

Phylogenetic tree object with polytomous branching, and known leaves and per node data

Phylo.SymmetricDiscreteTrait — Type

SymmetricDiscreteTrait{T <: AbstractTree, E <: Enum}

The simplest possible discrete trait evolved on a phylogenetic tree. This is a Sampleable type, so a random trait can be created using rand(sdt). The trait to be evolved must be an Enum (generally created using @enum), and is the second argument to the constructor:

function DiscreteTrait(tree::AbstractTree, ttype::Type{<:Enum}, transition_rate::Number, trait::String = "ttype")

The transition matrix holds transition rates from row to column (so row sums must be zero), and the transition probabilities in a branch are calculated as exp(transition_matrix .* branch_length).

Phylo.Ultrametric — Type

Ultrametric{T <: AbstractTree,
            SAMP <: Sampleable,
            LenUnits <: Number}(n::Int,
                                sampleable::SAMP = Exponential())
Ultrametric{T <: AbstractTree,
            SAMP <: Sampleable,
            LenUnits <: Number}(tiplabels::Vector{String},
                                sampleable::SAMP = Exponential())

The sampler for ultrametric phylogenetic trees of size n or with tip labels tiplabels. Generate random trees by calling rand().

LightGraphs.dst — Function

dst(tree::AbstractTree, branch)

Return the destination node for this branch.

LightGraphs.indegree — Function

indegree(tree::AbstractTree, node)

Return in degree of node.

LightGraphs.outdegree — Function

outdegree(tree::AbstractTree, node)

Return out degree of node.

LightGraphs.src — Function

src(tree::AbstractTree, branch)

Return the source node for this branch.

Phylo.branchdatatype — Method

branchdatatype(::Type{<: AbstractTree})

retrieve the branch info type of a tree.

Phylo.branchdims — Method

branchdims(::Type{<: AbstractTree})

retrieve the dimensions of the branch lengths for the tree.

Phylo.branchfilter — Method

branchfilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the branches of any tree, where the AbstractBranch is filtered by the function filterfn.

Phylo.branchfuture — Function

branchfuture(tree::AbstractTree, node)

Find the branches between a node on a tree and its leaves

Phylo.branchhistory — Function

branchhistory(tree::AbstractTree, node)

Find the branch route between a node on a tree and its root

Phylo.branchiter — Method

branchiter(tree::AbstractTree)

Returns an iterator over the branches of any tree.

Phylo.branchnamefilter — Method

branchnamefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the names of the branches of any tree, where the AbstractBranch is filtered by the function filterfn.

Phylo.branchnameiter — Method

branchnameiter(tree::AbstractTree)

Returns an iterator over the names of branches of any tree.

Phylo.branchnametype — Method

branchnametype(::AbstractTree)

Returns type of branch names from a branch type.

Phylo.branchroute — Function

branchroute(tree::AbstractTree, node1, node2)

Find the branch route between two nodes on a tree

Phylo.branchtype — Method

branchtype(::Type{AbstractTree})

Returns type of branches from a tree type.

Phylo.createbranch! — Function

createbranch!(tree::AbstractTree, src, dst[, len::Number];
              data)

Add a branch from src to dst on tree with optional length and data. source and destination can be either nodes or nodenames.

Phylo.createnode! — Method

createnode!(tree::AbstractTree[, nodename]; data)

Create a node on a tree with optional node info.

Phylo.createnodes! — Function

createnodes!(tree::AbstractTree, count::Integer)
createnodes!(tree::AbstractTree, nodenames)
createnodes!(tree::AbstractTree, nodedict)

Add a number of nodes, a vector with given names, or a Dict with node names and associated node info to a tree.

Phylo.deletebranch! — Function

deletebranch!(tree::AbstractTree, branch)
deletebranch!(tree::AbstractTree, src, dst)

Delete the branch branch from tree, or branch connecting src node to dst node.

Phylo.deletenode! — Function

deletenode!(tree::AbstractTree, node)

Delete a node (or a name) from a tree

Phylo.distance — Method

distance(tree::AbstractTree, node1, node2)

Distance between two nodes on a tree

Phylo.distances — Method

distances(tree::AbstractTree)

Pairwise distances between all leaf nodes on a tree

Phylo.droptips! — Function

droptips!(tree::AbstractTree{OneTree}, tips)

Function to drop tips from a phylogenetic tree tree, which are found in the vector of tips or tip names, tips.

Phylo.getancestors — Function

getancestors(tree::AbstractTree, node)

Return the name of all of the nodes that are ancestral to this node.

Phylo.getbranch — Function

getbranch(tree::AbstractTree, branch)
getbranch(tree::AbstractTree, source, dest)

Returns a branch from a tree by name or by source and destination node.

Phylo.getbranchdata — Function

getbranchdata(::AbstractTree, label)

retrieve the branch data for a leaf of the tree.

Phylo.getbranches — Function

getbranches(::AbstractTree)

Returns the vector of branches of a single tree, or a Dict of vectors of branches for multiple trees.

Phylo.getbranchname — Function

getbranchname(::AbstractTree, branch)
getbranchname(branch)

Returns the branch name associated with a branch from a tree. For some branch types, it will be able to extract the branch name without reference to the tree.

Phylo.getbranchnames — Function

getbranchnames(tree::AbstractTree)

Return a vector of branch names of a single tree, or a Dict of vectors of branch names for multiple trees.

Phylo.getchildren — Function

getchildren(tree::AbstractTree, node)

Return the [name(s) of] the child node(s) for this node [name].

Phylo.getdescendants — Function

getdescendants(tree::AbstractTree, node)

Return the names of all of the nodes that descend from this node.

Phylo.getheight — Function

getheight(tree::AbstractTree, node)

Return the height of the node.

Phylo.getinbound — Function

getinbound(tree::AbstractTree, node)

return the inbound branch to this node (returns name for node name, branch for node).

Phylo.getleafinfo — Function

getleafinfo(::AbstractTree[, label])

retrieve the leaf info for a leaf of the tree.

Phylo.getleafnames — Function

getleafnames(::AbstractTree[, ::TraversalOrder])

Retrieve the leaf names from the tree (in some specific order).

Phylo.getleaves — Function

getleaves(::AbstractTree[, ::TraversalOrder])

Retrieve the leaves from the tree.

Phylo.getlength — Function

getlength(tree::AbstractTree, branch)

Return the length of this branch.

Phylo.getnode — Method

getnode(tree::AbstractTree, nodename)

Returns a node from a tree.

Phylo.getnodedata — Function

getnodedata(::AbstractTree, node)

retrieve the node data for a node of the tree.

Phylo.getnodename — Function

getnodename(::AbstractTree, node)

Returns the node name associated with a node from a tree. For some node types, it will be able to extract the node name without reference to the tree.

Phylo.getnodenames — Function

getnodenames(::AbstractTree[, ::TraversalOrder])

Return a vector of node names of a single tree (identified by id for a ManyTrees tree), or a Dict of vectors of node names for multiple trees.

Phylo.getnodes — Function

getnodes(::AbstractTree[, ::TraversalOrder])

Returns the vector of nodes of a single tree, or a Dict of vectors of nodes for multiple trees.

Phylo.getoutbounds — Function

getoutbounds(tree::AbstractTree, nodename)

Return the names of the outbound branches from this node.

Phylo.getparent — Function

getparent(tree::AbstractTree, node)

Return [the name of] the parent node for this node [name]. Second method may not be implemented for some node types.

Phylo.getroot — Function

getroot(::AbstractTree)

Returns the root of a single tree (must be only one tree for a ManyTrees tree).

Phylo.getrootheight — Method

getrootheight(tree::AbstractTree)

Get the tree's root height.

Phylo.getroots — Function

getroots(::AbstractTree)
getroots(::AbstractTree, id)

Returns a vector containing the root(s) of a single (OneTree) tree or a set of (ManyTrees) trees.

Phylo.gettreeinfo — Function

gettreeinfo(tree::AbstractTree)
gettreeinfo(tree::AbstractTree, treename)

Returns the info data associated with the tree(s).

Phylo.gettreename — Method

gettreename(tree::AbstractTree)

Returns the name of the single tree.

Phylo.gettreenames — Function

gettreenames(tree::AbstractTree)

Returns the names of the trees.

Phylo.gettrees — Method

gettrees(tree::AbstractTree)

Returns an iterable of trees.

Phylo.hasbranch — Function

hasbranch(tree::AbstractTree, branch)
hasbranch(tree::AbstractTree, source, dest)

Does tree have a branch branch or a branch from source to dest?

Phylo.hasheight — Function

hasheight(tree::AbstractTree, node)

Does the node have a height defined?

Phylo.hasinbound — Function

hasinbound(tree::AbstractTree, node)

Does the node have an inbound connection?

Phylo.hasinboundspace — Function

hasinboundspace(tree::AbstractTree, node)

Does the node have space for an inbound connection?

Phylo.hasnode — Method

hasnode(tree::AbstractTree, node)

Returns whether a tree has a given node (or node name) or not.

Phylo.hasoutboundspace — Function

hasoutboundspace(tree::AbstractTree, node)

Does the node have space for an[other] outbound connection?

Phylo.hasrootheight — Function

hasrootheight(tree::AbstractTree)

Does the tree have an explicit root height

Phylo.heightstoroot — Method

heights(tree::AbstractTree)

Height of all of the leaves of the tree above the root

Phylo.heighttoroot — Method

height(tree::AbstractTree, node)

Height of a node of the tree above the root

Phylo.isinternal — Function

isinternal(tree::AbstractTree, node)

Is the node (referenced by name or node object) internal to the tree (neither root nor leaf)?

Phylo.isleaf — Function

isleaf(tree::AbstractTree, node)

Is the node (referenced by name or node object) a leaf of the tree?

Phylo.isroot — Function

isroot(tree::AbstractTree, node)

Is the node (referenced by name or node object) a root of the tree?

Phylo.isunattached — Function

isunattached(tree::AbstractTree, node)

Is the node (referenced by name or node object) unattached (i.e. not connected to other nodes)?

Phylo.keeptips! — Function

keeptips!(tree::AbstractTree{OneTree}, tips)

Function to keep only the tips in a phylogenetic tree, tree, that are found in the vector of tips or tip names, tips.

Phylo.leafinfotype — Method

leafinfotype(::Type{<: AbstractTree})

retrieve the leaf info type of a tree.

Phylo.nbranches — Function

nbranches(::AbstractTree)

Returns the number of branches of a single tree, or a Dict of numbers of branches for multiple trees.

Phylo.nleaves — Method

nleaves(::AbstractTree)

Returns the number of leaves (tips) in a tree.

Phylo.nnodes — Function

nnodes(::AbstractTree)

Returns the number of nodes of a single tree, or a Dict of numbers of nodes for multiple trees.

Phylo.nodedatatype — Method

nodedatatype(::Type{<: AbstractTree})

retrieve the node info type of a tree.

Phylo.nodefilter — Method

nodefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the nodes of any tree, where the AbstractNode is filtered by the function filterfn.

Phylo.nodefuture — Function

nodefuture(tree::AbstractTree, node)

Find the nodes between a node on a tree and its leaves

Phylo.nodehistory — Function

nodehistory(tree::AbstractTree, node)

Find the node route between a node on a tree and its root

Phylo.nodeiter — Method

nodeiter(tree::AbstractTree)

Returns an iterator over the nodes of any tree.

Phylo.nodenamefilter — Method

nodenamefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the nodenames of any tree, where the AbstractNode itself is filtered by the function filterfn.

Phylo.nodenameiter — Method

nodenameiter(tree::AbstractTree)

Returns an iterator over the names of the nodes of any tree.

Phylo.nodenametype — Method

nodenametype(::Type{AbstractTree})
nodenametype(::Type{AbstractNode})
nodenametype(::Type{AbstractBranch})

Returns type of node names from a tree type.

Phylo.noderoute — Function

noderoute(tree::AbstractTree, node1, node2)

Find the node route between two nodes on a tree

Phylo.nodetype — Method

nodetype(::Type{AbstractTree})

Returns type of nodes from a tree type.

Phylo.nroots — Function

nroots(::AbstractTree)

Returns the number of roots in a tree. For OneTree types, Unrooted trees will return 0, OneRoot trees should return 1, and manyroots tree (ones with multiple subtrees) will return the number of subtrees. ManyTrees types will return a Dict of counts of the number of roots for each tree in the set.

Phylo.ntrees — Method

ntrees(tree::AbstractTree)

Returns the number of trees in a tree object, 1 for a OneTree tree type, and the count of trees for a ManyTrees type.

Phylo.roottype — Method

roottype(::Type{AbstractTree})
roottype(::Type{AbstractNode})
roottype(::Type{AbstractBranch})

Returns root type from a tree type.

Phylo.setbranchdata! — Function

setbranchdata!(::AbstractTree, branch, label, value)
setbranchdata!(::AbstractTree, branch, data)

Set the branch data for a branch of the tree.

Phylo.setheight! — Function

setheight!(tree::AbstractTree, nodename, height)

Set the height of the node.

Phylo.setleafinfo! — Method

setleafinfo!(::AbstractTree, table)

Set the leaf info for the leaves of the tree.

Phylo.setnodedata! — Function

setnodedata!(::AbstractTree, node, label, value)
setnodedata!(::AbstractTree, node, data)

Set the node data for a node of the tree.

Phylo.setrootheight! — Function

setrootheight!(tree::AbstractTree, height)

Set the tree's root height.

Phylo.traversal — Function

traversal(::AbstractTree, ::TraversalOrder)
traversal(::AbstractTree, ::TraversalOrder, init)

Return an iterable object for a tree containing nodes in given order - preorder, inorder, postorder or breadthfirst - optionally starting from init.

Phylo.treenametype — Method

treenametype(::Type{AbstractTree})

Returns the name type for a tree type.

Phylo.validate! — Method

validate!(tree::AbstractTree)

Validate the tree by making sure that it is connected up correctly.

Phylo.API

Phylo.Phylo

Phylo.BinaryNode

Phylo.BinaryTree

Phylo.Branch

Phylo.BrownianTrait

Phylo.DiscreteTrait

Phylo.NamedBinaryTree

Phylo.NamedTree

Phylo.Node

Phylo.Nonultrametric

Phylo.PolytomousTree

Phylo.SymmetricDiscreteTrait

Phylo.Ultrametric

LightGraphs.dst

LightGraphs.indegree

LightGraphs.outdegree

LightGraphs.src

Phylo.API._addconnection!

Phylo.API._addinbound!

Phylo.API._addoutbound!

Phylo.API._branchdatatype

Phylo.API._branchdims

Phylo.API._clearrootheight!

Phylo.API._createbranch!

Phylo.API._createnode!

Phylo.API._degree

Phylo.API._deletebranch!

Phylo.API._deletenode!

Phylo.API._dst

Phylo.API._getbranch

Phylo.API._getbranches

Phylo.API._getbranchname

Phylo.API._getbranchnames

Phylo.API._getchildren

Phylo.API._getconnections

Phylo.API._getheight

Phylo.API._getinbound

Phylo.API._getleafnames

Phylo.API._getleaves

Phylo.API._getlength

Phylo.API._getnode

Phylo.API._getnodenames

Phylo.API._getnodes

Phylo.API._getoutbounds

Phylo.API._getparent

Phylo.API._getroot

Phylo.API._getrootheight

Phylo.API._getroots

Phylo.API._gettree

Phylo.API._gettreeinfo

Phylo.API._gettreename

Phylo.API._gettreenames

Phylo.API._gettrees

Phylo.API._hasbranch

Phylo.API._hasheight

Phylo.API._hasinbound

Phylo.API._hasinboundspace

Phylo.API._hasnode

Phylo.API._hasoutboundspace

Phylo.API._hasrootheight

Phylo.API._indegree

Phylo.API._isinternal

Phylo.API._isleaf

Phylo.API._isroot

Phylo.API._isunattached

Phylo.API._leafinfotype

Phylo.API._nbranches

Phylo.API._newbranchlabel

Phylo.API._newnodelabel

Phylo.API._nleaves

Phylo.API._nnodes

Phylo.API._nodedatatype

Phylo.API._nroots

Phylo.API._ntrees

Phylo.API._outdegree

Phylo.API._preferbranchobjects

Phylo.API._prefernodeobjects

Phylo.API._removeconnection!

Phylo.API._removeinbound!

Phylo.API._removeoutbound!

Phylo.API._resetleaves!

Phylo.API._setheight!

Phylo.API._setrootheight!

Phylo.API._src

Phylo.API._traversal

Phylo.API._treenametype

Phylo.API._validate!

Phylo.branchdatatype

Phylo.branchdims

Phylo.branchfilter

Phylo.branchfuture

Phylo.branchhistory

Phylo.branchiter

Phylo.branchnamefilter

Phylo.branchnameiter

Phylo.branchnametype

Phylo.branchroute

Phylo.branchtype

Phylo.createbranch!

Phylo.createnode!

Phylo.createnodes!

Phylo.deletebranch!

Phylo.deletenode!

Phylo.distance

Phylo.distances

Phylo.droptips!

Phylo.getancestors

Phylo.getbranch

Phylo.getbranchdata

Phylo.getbranches

Phylo.getbranchname

Phylo.getbranchnames

Phylo.getchildren

Phylo.getdescendants

Phylo.getheight

Phylo.getinbound

Phylo.getleafinfo

Phylo.getleafnames

Phylo.getleaves

Phylo.getlength

Phylo.getnode

Phylo.getnodedata

Phylo.getnodename

Phylo.getnodenames

Phylo.getnodes

Phylo.getoutbounds

Phylo.getparent

Phylo.getroot

Phylo.getrootheight

Phylo.getroots

Phylo.gettreeinfo

Phylo.gettreename

Phylo.gettreenames

Phylo.gettrees

Phylo.hasbranch

Phylo.hasheight

Phylo.hasinbound

Phylo.hasinboundspace

Phylo.hasnode

Phylo.hasoutboundspace

Phylo.hasrootheight

Phylo.heightstoroot

Phylo.heighttoroot

Phylo.isinternal

Phylo.isleaf

Phylo.isroot

Phylo.isunattached

Phylo.keeptips!

Phylo.leafinfotype

Phylo.nbranches

Phylo.nleaves

Phylo.nnodes

Phylo.nodedatatype

Phylo.nodefilter

Phylo.nodefuture

Phylo.nodehistory

Phylo.nodeiter

Phylo.nodenamefilter

Phylo.nodenameiter

Phylo.nodenametype

Phylo.noderoute

Phylo.nodetype

Phylo.nroots

Phylo.ntrees

Phylo.roottype

Phylo.setbranchdata!

Phylo.setheight!

Phylo.setleafinfo!

Phylo.setnodedata!

Phylo.setrootheight!

Phylo.traversal

Phylo.treenametype

Phylo.validate!