Building and manipulating trees

TODO: examples here are needed

Phylo.createbranch! — Function

createbranch!(tree::AbstractTree, src, dst[, len::Number];
              data)

Add a branch from src to dst on tree with optional length and data. source and destination can be either nodes or nodenames.

Phylo.deletebranch! — Function

deletebranch!(tree::AbstractTree, branch)
deletebranch!(tree::AbstractTree, src, dst)

Delete the branch branch from tree, or branch connecting src node to dst node.

Phylo.createnode! — Function

createnode!(tree::AbstractTree[, nodename]; data)

Create a node on a tree with optional node info.

Phylo.createnodes! — Function

createnodes!(tree::AbstractTree, count::Integer)
createnodes!(tree::AbstractTree, nodenames)
createnodes!(tree::AbstractTree, nodedict)

Add a number of nodes, a vector with given names, or a Dict with node names and associated node info to a tree.

Phylo.deletenode! — Function

deletenode!(tree::AbstractTree, node)

Delete a node (or a name) from a tree

Phylo.droptips! — Function

droptips!(tree::AbstractTree{OneTree}, tips; keep = false)

Function to drop tips from a phylogenetic tree tree, which are found in the vector of tips or tip names, tips. keep determines whether to keep internal and root nodes that now only have one child (default is false). Internal nodes with no children will always be removed.

Phylo.keeptips! — Function

keeptips!(tree::AbstractTree{OneTree}, tips; keep = false)

Function to keep only the tips in a phylogenetic tree, tree, that are found in the vector of tips or tip names, tips. keep determines whether to keep internal and root nodes that now only have one child (default is false). Internal nodes with no children will always be removed.