Routes along the tree

Phylo.branchhistory — Function

branchhistory(tree::AbstractTree, node)

Find the branch route between a node on a tree and its root

Phylo.branchfuture — Function

branchfuture(tree::AbstractTree, node)

Find the branches between a node on a tree and its leaves

Phylo.branchroute — Function

branchroute(tree::AbstractTree, node1, node2)

Find the branch route between two nodes on a tree

Phylo.nodehistory — Function

nodehistory(tree::AbstractTree, node)

Find the node route between a node on a tree and its root

Phylo.nodefuture — Function

nodefuture(tree::AbstractTree, node)

Find the nodes between a node on a tree and its leaves

Phylo.noderoute — Function

noderoute(tree::AbstractTree, node1, node2)

Find the node route between two nodes on a tree