Traversing and iterating over trees

The code also provides iterators, and filtered iterators over the branches, nodes, branchnames and nodenames of a tree (using the random tree from Creating and writing phylogenies)

using Phylo
nu = Nonultrametric(5);
tree = rand(nu)

collect(nodeiter(tree))

9-element Vector{LinkNode{OneRoot, String, Dict{String, Any}, LinkBranch{OneRoot, String, Dict{String, Any}, Float64}}}:
 LinkNode tip 1, a tip of the tree with an incoming connection (branch 12).
 LinkNode tip 2, a tip of the tree with an incoming connection (branch 9).
 LinkNode tip 3, a tip of the tree with an incoming connection (branch 15).
 LinkNode tip 4, a tip of the tree with an incoming connection (branch 10).
 LinkNode tip 5, a tip of the tree with an incoming connection (branch 13).
 LinkNode Node 6, an internal node with 1 inbound and 2 outbound connections (branches 11 and 9, 10)
 LinkNode Node 7, an internal node with 1 inbound and 2 outbound connections (branches 14 and 11, 12)
 LinkNode Node 8, an internal node with 1 inbound and 2 outbound connections (branches 16 and 13, 14)
 LinkNode Node 9, a root node with 2 outbound connections (branches 15, 16)

collect(nodenamefilter(isroot, tree))

1-element Vector{String}:
 "Node 9"

TreeSets are iterators themselves

trees = rand(nu, ["Tree 1", "Tree 2"])
collect(trees)

2-element Vector{LinkTree{OneRoot, String, LinkNode{OneRoot, String, Dict{String, Any}, LinkBranch{OneRoot, String, Dict{String, Any}, Float64}}, LinkBranch{OneRoot, String, Dict{String, Any}, Float64}, Dict{String, Any}}}:
 LinkTree{OneRoot, String, LinkNode{OneRoot, String, Dict{String, Any}, LinkBranch{OneRoot, String, Dict{String, Any}, Float64}}, LinkBranch{OneRoot, String, Dict{String, Any}, Float64}, Dict{String, Any}} with 5 tips and 1 root. Leaf names are tip 4, tip 1, tip 2, tip 5 and tip 3

9 nodes: [LinkNode Node 9, a root node with 2 outbound connections (branches 15, 16), LinkNode Node 8, an internal node with 1 inbound and 2 outbound connections (branches 15 and 13, 14), LinkNode Node 7, an internal node with 1 inbound and 2 outbound connections (branches 13 and 11, 12), LinkNode Node 6, an internal node with 1 inbound and 2 outbound connections (branches 11 and 9, 10), LinkNode tip 4, a tip of the tree with an incoming connection (branch 9)., LinkNode tip 1, a tip of the tree with an incoming connection (branch 10)., LinkNode tip 2, a tip of the tree with an incoming connection (branch 12)., LinkNode tip 5, a tip of the tree with an incoming connection (branch 14). and LinkNode tip 3, a tip of the tree with an incoming connection (branch 16).]

8 branches: [LinkBranch 9, from node Node 6 to node tip 4 (length 0.04776447042173523)., LinkBranch 10, from node Node 6 to node tip 1 (length 0.20565390580157905)., LinkBranch 11, from node Node 7 to node Node 6 (length 0.35108435266919935)., LinkBranch 12, from node Node 7 to node tip 2 (length 1.0533532225396265)., LinkBranch 13, from node Node 8 to node Node 7 (length 0.6436013178516)., LinkBranch 14, from node Node 8 to node tip 5 (length 0.8862004691715629)., LinkBranch 15, from node Node 9 to node Node 8 (length 0.041366871754222524). and LinkBranch 16, from node Node 9 to node tip 3 (length 0.008587786155407671).]

Node records: "Node 9" => Dict{String, Any}() ... "tip 3" => Dict{String, Any}()

 LinkTree{OneRoot, String, LinkNode{OneRoot, String, Dict{String, Any}, LinkBranch{OneRoot, String, Dict{String, Any}, Float64}}, LinkBranch{OneRoot, String, Dict{String, Any}, Float64}, Dict{String, Any}} with 5 tips and 1 root. Leaf names are tip 4, tip 1, tip 2, tip 5 and tip 3

9 nodes: [LinkNode Node 9, a root node with 2 outbound connections (branches 15, 16), LinkNode Node 8, an internal node with 1 inbound and 2 outbound connections (branches 15 and 13, 14), LinkNode Node 7, an internal node with 1 inbound and 2 outbound connections (branches 13 and 11, 12), LinkNode Node 6, an internal node with 1 inbound and 2 outbound connections (branches 11 and 9, 10), LinkNode tip 4, a tip of the tree with an incoming connection (branch 9)., LinkNode tip 1, a tip of the tree with an incoming connection (branch 10)., LinkNode tip 2, a tip of the tree with an incoming connection (branch 12)., LinkNode tip 5, a tip of the tree with an incoming connection (branch 14). and LinkNode tip 3, a tip of the tree with an incoming connection (branch 16).]

8 branches: [LinkBranch 9, from node Node 6 to node tip 4 (length 0.04776447042173523)., LinkBranch 10, from node Node 6 to node tip 1 (length 0.20565390580157905)., LinkBranch 11, from node Node 7 to node Node 6 (length 0.35108435266919935)., LinkBranch 12, from node Node 7 to node tip 2 (length 1.0533532225396265)., LinkBranch 13, from node Node 8 to node Node 7 (length 0.6436013178516)., LinkBranch 14, from node Node 8 to node tip 5 (length 0.8862004691715629)., LinkBranch 15, from node Node 9 to node Node 8 (length 0.041366871754222524). and LinkBranch 16, from node Node 9 to node tip 3 (length 0.008587786155407671).]

Node records: "Node 9" => Dict{String, Any}() ... "tip 3" => Dict{String, Any}()

Phylo.nodeiter — Function

nodeiter(tree::AbstractTree)

Returns an iterator over the nodes of any tree.

Phylo.nodefilter — Function

nodefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the nodes of any tree, where the AbstractNode is filtered by the function filterfn.

Phylo.nodenameiter — Function

nodenameiter(tree::AbstractTree)

Returns an iterator over the names of the nodes of any tree.

Phylo.nodenamefilter — Function

nodenamefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the nodenames of any tree, where the AbstractNode itself is filtered by the function filterfn.

Phylo.branchiter — Function

branchiter(tree::AbstractTree)

Returns an iterator over the branches of any tree.

Phylo.branchfilter — Function

branchfilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the branches of any tree, where the AbstractBranch is filtered by the function filterfn.

Phylo.branchnameiter — Function

branchnameiter(tree::AbstractTree)

Returns an iterator over the names of branches of any tree.

Phylo.branchnamefilter — Function

branchnamefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the names of the branches of any tree, where the AbstractBranch is filtered by the function filterfn.

Phylo.traversal — Function

traversal(::AbstractTree, ::TraversalOrder)
traversal(::AbstractTree, ::TraversalOrder, init)

Return an iterable object for a tree containing nodes in given order - preorder, inorder, postorder or breadthfirst - optionally starting from init.