Data types

Tree sets

Tree types

This package offers a number of different types of tree, each optimised for a specific usage

Node types

Branch types

Phylo.BinaryTree — Type

BinaryTree

Binary phylogenetic tree object with known leaves and per node data

Phylo.NamedBinaryTree — Type

NamedBinaryTree

Binary phylogenetic tree object with known leaves

Phylo.BinaryNode — Type

BinaryNode{B}(AbstractVector{B}, AbstractVector{B}) <: AbstractNode

A node of strict binary phylogenetic tree

Phylo.Node — Type

Node{RT, NL, T}(AbstractVector{T}, AbstractVector{T}) <: AbstractNode

A node of potentially polytomous phylogenetic tree

Phylo.NamedTree — Type

NamedPolytomousTree

Polytomous phylogenetic tree object with known leaves

Phylo.PolytomousTree — Type

PolytomousTree

Phylogenetic tree object with polytomous branching, and known leaves and per node data

Phylo.NamedPolytomousTree — Type

PolytomousTree

Phylogenetic tree object with polytomous branching, and known leaves and per node data

Missing docstring.

Missing docstring for LinkTree. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LinkBranch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LinkNode. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RootedTree. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ManyRootTree. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UnrootedTree. Check Documenter's build log for details.

Missing docstring.

Missing docstring for TreeSet. Check Documenter's build log for details.