Traversing and iterating over trees

The code also provides iterators, and filtered iterators over the branches, nodes, branchnames and nodenames of a tree (using the random tree from Creating phylogenies)

using Phylo
nu = Nonultrametric(5);
tree = rand(nu)

collect(nodeiter(tree))

9-element Vector{LinkNode{OneRoot, String, Dict{String, Any}, LinkBranch{OneRoot, String, Dict{String, Any}, Float64}}}:
 LinkNode tip 1, a tip of the tree with an incoming connection (branch 10).
 LinkNode tip 2, a tip of the tree with an incoming connection (branch 12).
 LinkNode tip 3, a tip of the tree with an incoming connection (branch 15).
 LinkNode tip 4, a tip of the tree with an incoming connection (branch 14).
 LinkNode tip 5, a tip of the tree with an incoming connection (branch 9).
 LinkNode Node 6, an internal node with 1 inbound and 2 outbound connections (branches 11 and 9, 10)
 LinkNode Node 7, an internal node with 1 inbound and 2 outbound connections (branches 13 and 11, 12)
 LinkNode Node 8, an internal node with 1 inbound and 2 outbound connections (branches 16 and 13, 14)
 LinkNode Node 9, a root node with 2 outbound connections (branches 15, 16)

collect(nodenamefilter(isroot, tree))

1-element Vector{String}:
 "Node 9"

TreeSets are iterators themselves

trees = rand(nu, ["Tree 1", "Tree 2"])
collect(trees)

2-element Vector{RootedTree}:
 RootedTree with 5 tips and 1 roots. Leaf names are tip 3, tip 1, tip 4, tip 2 and tip 5
9 nodes:
[root Node 9] --> (branch branch 15)
              --> (branch branch 16)
(branch 15) --> [leaf tip 3]
(branch 16) --> [internal Node 8] --> (branch 13)
                                  --> (branch 14)
(branch 13) --> [internal Node 7] --> (branch 11)
                                  --> (branch 12)
(branch 11) --> [leaf tip 1]
(branch 12) --> [leaf tip 4]
(branch 14) --> [internal Node 6] --> (branch 9)
                                  --> (branch 10)
(branch 9) --> [leaf tip 2]
(branch 10) --> [leaf tip 5]
8 branches:
["Node 6"] --> (0.4910376733799104 length branch 9) --> ["tip 2"]
["Node 6"] --> (0.3371910766649014 length branch 10) --> ["tip 5"]
["Node 7"] --> (0.12957867107385954 length branch 11) --> ["tip 1"]
["Node 7"] --> (0.628614436791121 length branch 12) --> ["tip 4"]
["Node 8"] --> (0.0022736084159711532 length branch 13) --> ["Node 7"]
["Node 8"] --> (0.026075492417705532 length branch 14) --> ["Node 6"]
["Node 9"] --> (0.3679043148174628 length branch 15) --> ["tip 3"]
["Node 9"] --> (0.747243906401348 length branch 16) --> ["Node 8"]
Node records:
Dict{String, Dict{String, Any}}("tip 2" => Dict(), "tip 5" => Dict(), "tip 4" => Dict(), "tip 1" => Dict(), "Node 9" => Dict(), "Node 8" => Dict(), "tip 3" => Dict(), "Node 7" => Dict(), "Node 6" => Dict())

 RootedTree with 5 tips and 1 roots. Leaf names are tip 3, tip 1, tip 4, tip 2 and tip 5
9 nodes:
[root Node 9] --> (branch branch 15)
              --> (branch branch 16)
(branch 15) --> [leaf tip 3]
(branch 16) --> [internal Node 8] --> (branch 13)
                                  --> (branch 14)
(branch 13) --> [internal Node 7] --> (branch 11)
                                  --> (branch 12)
(branch 11) --> [leaf tip 1]
(branch 12) --> [leaf tip 4]
(branch 14) --> [internal Node 6] --> (branch 9)
                                  --> (branch 10)
(branch 9) --> [leaf tip 2]
(branch 10) --> [leaf tip 5]
8 branches:
["Node 6"] --> (0.4910376733799104 length branch 9) --> ["tip 2"]
["Node 6"] --> (0.3371910766649014 length branch 10) --> ["tip 5"]
["Node 7"] --> (0.12957867107385954 length branch 11) --> ["tip 1"]
["Node 7"] --> (0.628614436791121 length branch 12) --> ["tip 4"]
["Node 8"] --> (0.0022736084159711532 length branch 13) --> ["Node 7"]
["Node 8"] --> (0.026075492417705532 length branch 14) --> ["Node 6"]
["Node 9"] --> (0.3679043148174628 length branch 15) --> ["tip 3"]
["Node 9"] --> (0.747243906401348 length branch 16) --> ["Node 8"]
Node records:
Dict{String, Dict{String, Any}}("tip 2" => Dict(), "tip 5" => Dict(), "tip 4" => Dict(), "tip 1" => Dict(), "Node 9" => Dict(), "Node 8" => Dict(), "tip 3" => Dict(), "Node 7" => Dict(), "Node 6" => Dict())

Phylo.nodeiter — Function

nodeiter(tree::AbstractTree)

Returns an iterator over the nodes of any tree.

Phylo.nodefilter — Function

nodefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the nodes of any tree, where the AbstractNode is filtered by the function filterfn.

Phylo.nodenameiter — Function

nodenameiter(tree::AbstractTree)

Returns an iterator over the names of the nodes of any tree.

Phylo.nodenamefilter — Function

nodenamefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the nodenames of any tree, where the AbstractNode itself is filtered by the function filterfn.

Phylo.branchiter — Function

branchiter(tree::AbstractTree)

Returns an iterator over the branches of any tree.

Phylo.branchfilter — Function

branchfilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the branches of any tree, where the AbstractBranch is filtered by the function filterfn.

Phylo.branchnameiter — Function

branchnameiter(tree::AbstractTree)

Returns an iterator over the names of branches of any tree.

Phylo.branchnamefilter — Function

branchnamefilter(filterfn::Function, tree::AbstractTree)

Returns an iterator over the names of the branches of any tree, where the AbstractBranch is filtered by the function filterfn.

Phylo.traversal — Function

traversal(::AbstractTree, ::TraversalOrder)
traversal(::AbstractTree, ::TraversalOrder, init)

Return an iterable object for a tree containing nodes in given order - preorder, inorder, postorder or breadthfirst - optionally starting from init.