Modules and data types

Modules

Phylo.Phylo — Module

Phylo package

The Phylo package provides some simple phylogenetics types (e.g. NamedTree) to interface to the Diversity package for measuring phylogenetic diversity. It also provides an interface to R for copying trees to and from that language and can read newick and nexus tree files (including TreeSets that contain multiple trees).

Finally it also provides a standard abstract interface to phylogenetic trees, by defining AbstractNode, AbstractBranch and AbstractTree supertypes, and methods to interface to them. It also provides (through the Phylo.API submodule) methods to (re)define to write your own phylogenetic type in a way that will interact cleanly with other phylogenetic packages.

Tree sets

Phylo.TreeSet — Type

TreeSet

A collection of trees with the same tips.

Tree types

This package offers a number of different types of tree, each optimised for a specific usage

Phylo.RecursiveTree — Type

struct RecursiveTree <: AbstractTree

A phylogenetic tree type containing RecursiveElts as both nodes and branches, allowing navigation of the tree using only the node and branch elements.

Phylo.LinkTree — Type

struct LinkTree <: AbstractTree

A phylogenetic tree type containing LinkNodes and LinkBranches

Phylo.NamedBinaryTree — Type

NamedBinaryTree

Binary phylogenetic tree object with known leaves

Phylo.BinaryTree — Type

BinaryTree

Binary phylogenetic tree object with known leaves and per node data

Phylo.NamedTree — Type

NamedPolytomousTree

Polytomous phylogenetic tree object with known leaves

Phylo.PolytomousTree — Type

PolytomousTree

Phylogenetic tree object with polytomous branching, and known leaves and per node data

Phylo.NamedPolytomousTree — Type

PolytomousTree

Phylogenetic tree object with polytomous branching, and known leaves and per node data

Node and Branch types

Phylo.RecursiveElt — Type

struct RecursiveElt <: AbstractElt

A type for branches or nodes in a RecursiveTree, allowing navigation of the tree without using the tree object itself.

Phylo.LinkNode — Type

struct LinkNode <: AbstractNode

A node type that is connected by LinkBranches in a LinkTree

Phylo.BinaryNode — Type

struct BinaryNode <: AbstractNode

A node of strict binary phylogenetic tree

Phylo.Node — Type

struct Node <: AbstractNode

A node of potentially polytomous phylogenetic tree

Phylo.LinkBranch — Type

struct LinkBranch <: AbstractBranch

A branch type that connects LinkNodes in a LinkTree

Phylo.Branch — Type

Branch

A branch connecting two AbstractNodes of a phylogenetic tree

Iterator types

Phylo.BranchIterator — Type

BranchIterator

The struct representing an iterator for branches of a phylogenetic tree

Phylo.BranchNameIterator — Type

BranchNameIterator

The struct representing an iterator for branchnames of a phylogenetic tree

Phylo.NodeIterator — Type

NodeIterator

The struct representing an iterator for nodes of a phylogenetic tree

Phylo.NodeNameIterator — Type

NodeNameIterator

The struct representing an iterator for nodenames of a phylogenetic tree